First pass frontend refactor
